extract form data logic to hook to simplify form component
